v0.000001

+++ Dojo/Model/DJMedia.cs
@@ -6,7 +6,7 @@
         public string mime_type { get; set; }
         public string link { get; set; }
         public string base64 { get; set; }
-        public int aspect_ratio { get; set; }
+        public float aspect_ratio { get; set; }
         public int width { get; set; }
         public int height { get; set; }
         public int order { get; set; }
